refactored constants to a single file
